Refactored CBLReadOnlyDocument to have a (not be a) dictionary
Having CBLReadOnlyDocument subclass CBLReadOnlyDictionary is problematic,
because its subclass CBLDocument needs to implement CBLDictionary but
can't make use of the base class anymore; instead it has to delegate to
a separate CBLDictionary object.

I've changed CBLReadOnlyDocument to delegate to a CBLReadOnlyDictionary
and just inherit from NSObject. This makes it easy for CBLDocument to
instantiate a CBLDictionary instead.

The downside is that CBLReadOnlyDocument has to implement all of the
CBLReadOnlyDictionary methods to delegate them; but this is canceled out
since CBLDictionary no longer has to override all those methods.